requirements.txt
aniso8601==7.0.0
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
cloudinary==1.29.0
Django==3.2
django-ckeditor==6.4.2
django-cloudinary-storage==0.3.0
django-filter==22.1
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
django-js-asset==2.0.0
graphene==2.1.9
graphene-django==2.15.0
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==19.9.0
idna==3.3
Pillow==9.2.0
promise==2.3
psycopg2-binary==2.9.5
PyJWT==1.7.1
pytz==2022.1
redis==3.2.1
requests==2.28.1
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
urllib3==1.26.11
